Initialized Repo: brought existing code id
Starting commit for personal package that is intended as a helper
package containing some useful structs. For now, will only work on the
Result class that is based on the rust enum type, but with changes to
adapt it better to my desired workflow. Future work is to have some
other monad-like classes
On branch main
Changes to be committed:
        new file:   LICENSE
        new file:   README.md
        new file:   old-or-deprecated/old-result-versions/MonadicList.py
        new file:   old-or-deprecated/old-result-versions/MonadicListOfResult.py
        new file:   old-or-deprecated/old-result-versions/OldResult.py
        new file:   old-or-deprecated/old-result-versions/anotherResult.py
        new file:   pyproject.toml
        new file:   src/__init__.py
        new file:   src/helper_structs/__init__.py
        new file:   src/helper_structs/__pycache__/__init__.cpython-312.pyc
        new file:   src/helper_structs/__pycache__/_types.cpython-312.pyc
        new file:   src/helper_structs/__pycache__/result.cpython-312.pyc
        new file:   src/helper_structs/_types.py
        new file:   src/helper_structs/result.py