Added new map method and bind method
Overwrote the previous map method which was a different specification
from the Rust Result map, so I made it more "compliant". At the same
time, extracted the contents of map into a new method bind, which is the
way that I had it structured a long way back. Also overrode the right
shift operator to chain bind calls.

Added docstring tests for more methods as well

On branch develop
Changes to be committed:
	modified:   src/helper_structs/__pycache__/_types.cpython-312.pyc
	modified:   src/helper_structs/__pycache__/result.cpython-312.pyc
	modified:   src/helper_structs/result.py